Major refactor to be more pydantic and dev-like. Still a WIP although 90% there
