+++ app/lock_manager.py
@@ -1,0 +1,3 @@
+import threading
+
+lock = threading.Lock()
